requirements.txt
pytesseract
Pillow
flask